# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IntelliAI 2.0 \u2014 Business Control Layer",
    "description": "REST API for the IntelliAI 2.0 IBM MQ migration control plane.\n\n**Team:** intelliAI2DotO  **Event:** Wells Fargo Hackathon 2026 \u2014 Phase 2",
    "contact": {
      "name": "intelliAI2DotO"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/topologies": {
      "get": {
        "tags": [
          "topology"
        ],
        "summary": "List all topologies",
        "operationId": "list_topologies_topologies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TopologyOut"
                  },
                  "type": "array",
                  "title": "Response List Topologies Topologies Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "topology"
        ],
        "summary": "Create a topology from a flow spec",
        "operationId": "create_topology_topologies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopologySpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopologyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/topologies/{topology_id}": {
      "get": {
        "tags": [
          "topology"
        ],
        "summary": "Get one topology with QM detail",
        "operationId": "get_topology_topologies__topology_id__get",
        "parameters": [
          {
            "name": "topology_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Topology Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopologyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/topologies/{topology_id}/applications": {
      "get": {
        "tags": [
          "topology"
        ],
        "summary": "Distinct applications in this topology",
        "operationId": "list_topology_applications_topologies__topology_id__applications_get",
        "parameters": [
          {
            "name": "topology_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Topology Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationOut"
                  },
                  "title": "Response List Topology Applications Topologies  Topology Id  Applications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audit": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Paged listing of audit log entries",
        "operationId": "list_audit_audit_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lamport clock value to fetch entries strictly less than",
              "title": "Cursor"
            },
            "description": "Lamport clock value to fetch entries strictly less than"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "app_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "App Id"
            }
          },
          {
            "name": "operation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AuditOperation"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operation"
            }
          },
          {
            "name": "correlation_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Correlation Id"
            }
          },
          {
            "name": "include_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Total"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audit/{lamport}": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Fetch one audit entry by Lamport timestamp",
        "operationId": "get_audit_entry_audit__lamport__get",
        "parameters": [
          {
            "name": "lamport",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lamport"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEntryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApplicationOut": {
        "properties": {
          "app_id": {
            "type": "string",
            "title": "App Id"
          },
          "app_name": {
            "type": "string",
            "title": "App Name"
          },
          "neighbourhood": {
            "type": "string",
            "title": "Neighbourhood"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "app_id",
          "app_name",
          "neighbourhood",
          "created_at"
        ],
        "title": "ApplicationOut"
      },
      "AuditEntryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lamport_clock": {
            "type": "integer",
            "title": "Lamport Clock"
          },
          "wall_clock": {
            "type": "string",
            "format": "date-time",
            "title": "Wall Clock"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "operation": {
            "$ref": "#/components/schemas/AuditOperation"
          },
          "app_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Id"
          },
          "qm_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qm Name"
          },
          "request_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Payload"
          },
          "response_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Payload"
          },
          "state_before": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Before"
          },
          "state_after": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State After"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "is_rollback": {
            "type": "boolean",
            "title": "Is Rollback"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lamport_clock",
          "wall_clock",
          "correlation_id",
          "actor",
          "operation",
          "app_id",
          "qm_name",
          "request_payload",
          "response_payload",
          "state_before",
          "state_after",
          "success",
          "error_message",
          "duration_ms",
          "is_rollback"
        ],
        "title": "AuditEntryOut"
      },
      "AuditOperation": {
        "type": "string",
        "enum": [
          "TOPOLOGY_CREATED",
          "TOPOLOGY_DELETED",
          "QM_DEPLOYED",
          "QM_READY",
          "QM_DELETED",
          "MQSC_DEFINE_QLOCAL",
          "MQSC_DEFINE_QREMOTE",
          "MQSC_DEFINE_QXMIT",
          "MQSC_DEFINE_CHANNEL_SDR",
          "MQSC_DEFINE_CHANNEL_RCVR",
          "MQSC_DEFINE_CHANNEL_SVRCONN",
          "MQSC_DEFINE_CHLAUTH",
          "MQSC_ALTER_QMGR",
          "MQSC_DELETE_QLOCAL",
          "MQSC_DELETE_QREMOTE",
          "MQSC_DELETE_QXMIT",
          "MQSC_DELETE_CHANNEL",
          "MQSC_START_CHANNEL",
          "MQSC_STOP_CHANNEL",
          "MIGRATION_PLANNED",
          "MIGRATION_STATE_TRANSITION",
          "MIGRATION_STEP_STARTED",
          "MIGRATION_STEP_COMPLETED",
          "MIGRATION_STEP_FAILED",
          "VALIDATION_RUN",
          "ROLLBACK_INITIATED",
          "ROLLBACK_STEP",
          "ROLLBACK_COMPLETED",
          "ROLLBACK_FAILED",
          "GUARDRAIL_REJECTED",
          "AGENT_INVOCATION"
        ],
        "title": "AuditOperation",
        "description": "Operation taxonomy on the audit log.\n\nGranular enough that the rollback engine can synthesize an inverse\nMQSC command from the operation alone, given the operation's payload."
      },
      "AuditPage": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AuditEntryOut"
            },
            "type": "array",
            "title": "Entries"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "next_cursor",
          "total_count"
        ],
        "title": "AuditPage"
      },
      "FlowSpec": {
        "properties": {
          "flow_type": {
            "type": "string",
            "enum": [
              "Local",
              "Remote"
            ],
            "title": "Flow Type"
          },
          "producer_app_id": {
            "type": "string",
            "title": "Producer App Id"
          },
          "producer_app_name": {
            "type": "string",
            "title": "Producer App Name"
          },
          "producer_neighbourhood": {
            "type": "string",
            "title": "Producer Neighbourhood"
          },
          "producer_queue_manager": {
            "type": "string",
            "title": "Producer Queue Manager"
          },
          "producer_queue_name": {
            "type": "string",
            "title": "Producer Queue Name"
          },
          "producer_queue_type": {
            "type": "string",
            "enum": [
              "Local",
              "Remote"
            ],
            "title": "Producer Queue Type"
          },
          "transmit_queue_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transmit Queue Name"
          },
          "channel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Name"
          },
          "consumer_app_id": {
            "type": "string",
            "title": "Consumer App Id"
          },
          "consumer_app_name": {
            "type": "string",
            "title": "Consumer App Name"
          },
          "consumer_neighnourhood": {
            "type": "string",
            "title": "Consumer Neighnourhood"
          },
          "consumer_queue_manager": {
            "type": "string",
            "title": "Consumer Queue Manager"
          },
          "consumer_queue_name": {
            "type": "string",
            "title": "Consumer Queue Name"
          },
          "consumer_queue_type": {
            "type": "string",
            "enum": [
              "Local",
              "Remote"
            ],
            "title": "Consumer Queue Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "flow_type",
          "producer_app_id",
          "producer_app_name",
          "producer_neighbourhood",
          "producer_queue_manager",
          "producer_queue_name",
          "producer_queue_type",
          "consumer_app_id",
          "consumer_app_name",
          "consumer_neighnourhood",
          "consumer_queue_manager",
          "consumer_queue_name",
          "consumer_queue_type"
        ],
        "title": "FlowSpec",
        "description": "One row from the source/target CSV \u2014 one application-pair flow.\n\nThe schema preserves the input CSV's structure exactly, including the\ntypo `consumer_neighnourhood` (extra 'n') \u2014 we don't silently fix\nupstream data."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueueManagerOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "qm_name": {
            "type": "string",
            "title": "Qm Name"
          },
          "pod_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pod Name"
          },
          "service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Name"
          },
          "listener_port": {
            "type": "integer",
            "title": "Listener Port"
          },
          "web_port": {
            "type": "integer",
            "title": "Web Port"
          },
          "dlq_name": {
            "type": "string",
            "title": "Dlq Name"
          },
          "deployed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployed At"
          },
          "is_ready": {
            "type": "boolean",
            "title": "Is Ready"
          }
        },
        "type": "object",
        "required": [
          "id",
          "qm_name",
          "pod_name",
          "service_name",
          "listener_port",
          "web_port",
          "dlq_name",
          "deployed_at",
          "is_ready"
        ],
        "title": "QueueManagerOut"
      },
      "TopologyKind": {
        "type": "string",
        "enum": [
          "SOURCE",
          "TARGET"
        ],
        "title": "TopologyKind",
        "description": "Source vs target topology \u2014 the two API layers per FAQ Q6/Q7."
      },
      "TopologyOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/TopologyKind"
          },
          "spec": {
            "additionalProperties": true,
            "type": "object",
            "title": "Spec"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "queue_managers": {
            "items": {
              "$ref": "#/components/schemas/QueueManagerOut"
            },
            "type": "array",
            "title": "Queue Managers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "spec",
          "created_at"
        ],
        "title": "TopologyOut"
      },
      "TopologySpec": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/TopologyKind"
          },
          "flows": {
            "items": {
              "$ref": "#/components/schemas/FlowSpec"
            },
            "type": "array",
            "minItems": 1,
            "title": "Flows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "kind",
          "flows"
        ],
        "title": "TopologySpec",
        "description": "The full topology spec \u2014 what gets POSTed to /topologies."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}